+++ src/AT2/OpenGL/GlRenderer.h
@@ -59,7 +59,7 @@ public:
 	void Shutdown() override;
 
 	void DispatchCompute(glm::uvec3 threadGroupSize) override;
-	void Draw(Primitives::Primitive type, long int first, long int count, int numInstances = 1, int baseVertex = 0) override;
+	void Draw(Primitives::Primitive type, size_t first, long int count, int numInstances = 1, int baseVertex = 0) override;
 
 	void SetViewport(const AABB2d& viewport) override;
 	void ClearBuffer(const glm::vec4& color) override;
